backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
python-multipart>=0.0.6
pydantic>=2.5.0
python-dotenv>=1.0.0
neo4j>=5.0.0
aiohttp>=3.9.0
certifi>=2024.0.0
